Feature/app 299 update family page concepts styling (#382)
* Add error handling so concepts dont crash

* Remove unnecessary code

* Add label & quote component

* Move duplicate code into processConcepts

* Refactor into new components

* Refactor concepts code into new panel
