+++ src/components/concepts/ConceptsHead.tsx
@@ -1,21 +1,26 @@
 import { ExternalLink } from "@components/ExternalLink";
+import { Heading } from "@components/typography/Heading";
+import { Label } from "@components/labels/Label";
+import { Quote } from "@components/typography/Quote";
 
 export const ConceptsHead = () => {
   return (
     <div className="mb-4">
       <div className="space-y-3">
         <div className="flex items-center gap-2">
-          <h2 className="text-base font-medium text-neutral-800">Climate concepts</h2>
-          <span className="bg-blue-600 text-white text-xs font-medium px-1.5 py-0.5 rounded-sm">Beta</span>
+          <Heading level={2} className="text-base font-medium text-neutral-800">
+            Climate concepts
+          </Heading>
+          <Label>Beta</Label>
         </div>
-        <div className="pl-4 border-l-2 border-blue-600">
-          <p className="text-sm text-neutral-600">
+        <Quote>
+          <>
             This feature automatically detects climate concepts in documents. Accuracy is not 100%.{" "}
             <ExternalLink url="https://climatepolicyradar.org/concepts" className="text-gray-600 underline">
               Learn more
             </ExternalLink>
-          </p>
-        </div>
+          </>
+        </Quote>
       </div>
     </div>
   );
