+++ src/components/documents/ConceptsDocumentViewer.tsx
@@ -19,6 +19,7 @@ import useSearch from "@hooks/useSearch";
 import { ConceptsHead } from "@components/concepts/ConceptsHead";
 import { HiOutlineDotsHorizontal } from "react-icons/hi";
 import { ConceptsPopover } from "@components/popover/ConceptsPopover";
+import { ConceptsPanel } from "@components/concepts/ConceptsPanel";
 
 type TProps = {
   initialQueryTerm?: string | string[];
@@ -161,14 +162,6 @@ export const ConceptsDocumentViewer = ({
     [onExactMatchChange]
   );
 
-  const handleConceptClick = useCallback(
-    (conceptLabel: string) => {
-      setState({ passageIndex: 0 });
-      onConceptClick?.(conceptLabel);
-    },
-    [onConceptClick]
-  );
-
   const handleClearSearch = useCallback(() => {
     setState({
       queryTerm: "",
@@ -258,75 +251,12 @@ export const ConceptsDocumentViewer = ({
                   )}
 
                   {selectedConcepts.length === 0 && !initialQueryTerm && (
-                    <div className="pb-4">
-                      <div className="mt-4 grow-0 shrink-0">
-                        <ConceptsHead></ConceptsHead>
-                      </div>
-
-                      {rootConcepts.map((rootConcept) => {
-                        const hasConceptsInRootConcept = concepts.filter((concept) => concept.subconcept_of.includes(rootConcept.wikibase_id));
-                        if (hasConceptsInRootConcept.length === 0) return null;
-                        return (
-                          <div key={rootConcept.wikibase_id} className="pt-6 pb-6 relative group">
-                            <div className="flex items-center gap-2">
-                              <p className="capitalize text-neutral-800 text-base font-medium leading-normal flex-grow">
-                                {rootConcept.preferred_label}
-                              </p>
-                              <div className="relative pr-3">
-                                <button
-                                  onClick={(e) => {
-                                    e.preventDefault();
-                                    setOpenPopoverIds(
-                                      openPopoverIds.includes(rootConcept.wikibase_id)
-                                        ? openPopoverIds.filter((id) => id !== rootConcept.wikibase_id)
-                                        : [...openPopoverIds, rootConcept.wikibase_id]
-                                    );
-                                  }}
-                                  className="text-neutral-500 flex items-center z-50"
-                                >
-                                  <HiOutlineDotsHorizontal className="text-xl group-hover:border-neutral-200 border-transparent border-1 rounded-full p-0.5" />
-                                </button>
-
-                                {openPopoverIds.includes(rootConcept.wikibase_id) && (
-                                  <div className="absolute z-50 top-full right-3 mt-2">
-                                    <ConceptsPopover
-                                      concept={rootConcept}
-                                      onClose={() => setOpenPopoverIds(openPopoverIds.filter((id) => id !== rootConcept.wikibase_id))}
-                                    />
-                                  </div>
-                                )}
-                              </div>
-                            </div>
-                            <ul className="flex flex-wrap gap-2 mt-4">
-                              {concepts
-                                .filter((concept) => concept.subconcept_of.includes(rootConcept.wikibase_id))
-                                .map((concept) => {
-                                  return (
-                                    <li key={concept.wikibase_id}>
-                                      <Link
-                                        className="capitalize hover:no-underline"
-                                        href="#"
-                                        onClick={(e) => {
-                                          e.preventDefault();
-                                          handleConceptClick?.(concept.preferred_label);
-                                        }}
-                                      >
-                                        <Button
-                                          color="clear-blue"
-                                          data-cy="view-document-viewer-concept"
-                                          extraClasses="capitalize flex items-center text-neutral-600 text-sm font-normal leading-tight"
-                                        >
-                                          {concept.preferred_label} {conceptCountsById[concept.wikibase_id]}
-                                        </Button>
-                                      </Link>
-                                    </li>
-                                  );
-                                })}
-                            </ul>
-                          </div>
-                        );
-                      })}
-                    </div>
+                    <ConceptsPanel
+                      rootConcepts={rootConcepts}
+                      concepts={concepts}
+                      conceptCountsById={conceptCountsById}
+                      onConceptClick={onConceptClick}
+                    ></ConceptsPanel>
                   )}
 
                   {selectedConcepts.length > 0 && (
