+++ src/components/popover/ConceptsPopover.tsx
@@ -1,4 +1,5 @@
 import { ExternalLink } from "@components/ExternalLink";
+import { Heading } from "@components/typography/Heading";
 import { TConcept } from "@types";
 import { getConceptStoreLink } from "@utils/getConceptStoreLink";
 
@@ -12,7 +13,9 @@ export const ConceptsPopover = ({ concept, onClose }: TProps) => {
     <div className="w-64 p-4 bg-white rounded-lg shadow-md border border-gray-200 flex items-start">
       <div className="space-y-6 w-full">
         <div className="space-y-1">
-          <h3 className="text-xs font-medium text-neutral-500">Description</h3>
+          <Heading level={3} className="text-xs font-medium text-neutral-500">
+            Description
+          </Heading>
           <p className="text-sm text-neutral-800">{concept?.description || "No description available"}</p>
         </div>
         <div className="space-y-1">
